refactor: adjust code and make small improvements.
