+++ src/SS.Game/Databases/SGUIDatabase.cs
@@ -1,9 +1,9 @@
-﻿using StardustSandbox.Game.Resources.GUISystem.Bundle.Hud;
-using StardustSandbox.Game.GUISystem;
+﻿using StardustSandbox.Game.GUISystem;
+using StardustSandbox.Game.Interfaces;
 using StardustSandbox.Game.Objects;
+using StardustSandbox.Game.Resources.GUISystem.Bundle.Hud;
 
 using System.Collections.Generic;
-using StardustSandbox.Game.Interfaces;
 
 namespace StardustSandbox.Game.Databases
 {
